chore: refactoring huge file
