+++ src/__init__.py
@@ -1,0 +1,15 @@
+# SerialBarcodeReaderTools package
+from .base_scanner import BaseScanner
+from .gm65_scanner import GM65Scanner
+from .m3yw_scanner import M3YWScanner
+from .main import main, detect_scanner
+from .utils import common_baud_rates
+
+__all__ = [
+    'BaseScanner',
+    'GM65Scanner', 
+    'M3YWScanner',
+    'main',
+    'detect_scanner',
+    'common_baud_rates'
+]
